fetch/session fix
