+++ .gitignore
@@ -1,3 +1,4 @@
+sessions
 
 # Created by https://www.gitignore.io/api/node,linux,macos,sublimetext,visualstudiocode
